Fix numpy array casting error
Newer versions of numpy resulted in array type cast issues caused by
1) Python treating stiffness/damping matrices as tuples (a syntactical
error on my part) and 2) numpy treating elements of arrays, e.g.,
velocity[1], as 1-D arrays of one element instead of extracting the
element as a float. This is remedied by accessing the item using
ndarray.item(), e.g., velocity[1].item(). This commit makes these
changes, allowing the program to be used with the latest versions of
numpy.